requirements/common.txt
Django==1.11.8
djangorestframework==3.8.2
django-filter==1.0.4